more main refactoring
