+++ src/http/mod.rs
@@ -1,5 +1,14 @@
+pub mod activity_map;
+pub mod api_key;
+pub mod audit_log;
 pub mod client;
 pub mod common;
 pub mod device;
 pub mod firmware;
 pub mod getter;
+
+pub mod prelude {
+    pub use crate::http::activity_map::get_activity_maps;
+    pub use crate::http::api_key::get_api_keys;
+    pub use crate::http::audit_log::get_audit_logs;
+}
